+++ packages/miniprogram-element/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * rich-text 的 nodes 支持 json 串
 * picker、picker-view 的 value 支持 json 串
+* 支持 swiper 的 snap-to-edge 属性
+* 修复 scroll-view 下拉刷新相关 bug
 
 ## 2.1.16