+++ packages/miniprogram-render/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * innerHTML 补充 behavior 输出
 * getElementsByTagName/querySelector/querySelectorAll 支持内置组件标签
+* select 标签在没有设置 value 的时候，取选中的 option 的 value 作为 select 的 value
+* 修复设置 innerHTML 为空串无法删除子节点的问题
 
 ## 2.1.7